Add simple example
